+++ src/components/Player/PlayerLocal/helpers.ts
@@ -1,0 +1,8 @@
+export const isPlaying = (player: HTMLVideoElement) => {
+  return (
+    player.currentTime > 0 &&
+    !player.paused &&
+    !player.ended &&
+    player.readyState > 2
+  );
+};